Tweak design of collapsible setting fieldsets
Chris Goffnett's report of overlapping content made me see something
I couldn't un-see. The arrows for content that gets collapsed were
hanging out in the margin. The previous commit looked fine on desktop
browsers but it looked awkward on mobile devices because the arrows
butted right up against the left side of the screen.

This commit moves the arrow in AND moves the border from below
the heading (and between the heading and the fieldset content) to
directly above the heading. As a result, a few margins had to be
adjusted to maintain spacing between groups of settings.

I hope we like this.