Fix premature stacking of two column layouts
This is a revised style. The original selector messed up all
layouts. Now this only adjusts the left margin of two column
pages WITH the small column on the right. Namely: The front
page and Bud's ENG 222 courses.

It takes great pains to leave three column layouts alone AND
two column layouts where the small column is on the left.